MANUAL: inlines Region and Subregion interfaces
