+++ libs/globetrotter/learn/util/src/index.ts
@@ -1,5 +1,3 @@
 export * from './lib/place.type';
 export * from './lib/quiz-type.enum';
-export * from './lib/region.interface';
 export * from './lib/shuffle';
-export * from './lib/subregion.interface';
