--- libs/globetrotter/learn/util/src/lib/region.interface.ts
@@ -1,6 +1,0 @@
-import { Subregion } from './subregion.interface';
-
-export interface Region {
-  name: string;
-  subregions: Subregion[];
-}
