--- libs/globetrotter/learn/util/src/lib/subregion.interface.ts
@@ -1,6 +1,0 @@
-import { Country } from '@atocha/globetrotter/shared/util';
-
-export interface Subregion {
-  name: string;
-  countries: Country[];
-}
